fixed bug when fetching ticks; minor refactoring of fetch.ts
