+++ protocol/ethereum/uniswapv3/wrapper/src/utils/fetch.ts
@@ -11,13 +11,18 @@ import {
   Args_fetchPoolFromTokens,
   FeeAmount,
   Ethereum_Module,
-  Subgraph_Module,
   Ethereum_Connection,
   Currency,
 } from "../wrap";
 import { createPool, getPoolAddress } from "../pool";
 import { _wrapToken } from "../token";
 import { getFeeAmountEnum } from "./enumUtils";
+import {
+  getSubgraphEndpoint,
+  SubgraphEndpoint,
+  subgraphQuery,
+} from "./subgraph";
+import { ethCallView, poolAbi, PoolImmutables, PoolState } from "./poolView";
 
 import { BigInt, JSON } from "@polywrap/wasm-as";
 
@@ -135,78 +140,7 @@ export function fetchPoolFromAddress(args: Args_fetchPoolFromAddress): Pool {
  * @param args.chainId the id of the chain to be queried
  */
 export function fetchTickList(args: Args_fetchTickList): Tick[] {
-  return fetchPoolTicksSubgraph(args.address, args.chainId);
-}
-
-class PoolImmutables {
-  token0: string;
-  token1: string;
-  fee: FeeAmount;
-}
-
-class PoolState {
-  liquidity: BigInt;
-  sqrtPriceX96: BigInt;
-  tick: i32;
-}
-
-class SubgraphEndpoint {
-  author: string;
-  name: string;
-}
-
-class QueryArgs {
-  subgraphAuthor: string;
-  subgraphName: string;
-  query: string;
-}
-
-function poolAbi(methodName: string): string {
-  if (methodName == "token0") {
-    return "function token0() external view returns (address)";
-  } else if (methodName == "token1") {
-    return "function token1() external view returns (address)";
-  } else if (methodName == "fee") {
-    return "function fee() external view returns (uint24)";
-  } else if (methodName == "liquidity") {
-    return "function liquidity() external view returns (uint128)";
-  } else if (methodName == "slot0") {
-    return "function slot0() external view returns (uint160 sqrtPriceX96, int24 tick, uint16 observationIndex, uint16 observationCardinality, uint16 observationCardinalityNext, uint8 feeProtocol, bool unlocked)";
-  } else if (methodName == "ticks") {
-    return "function ticks(int24 tick) external view returns (uint128 liquidityGross, int128 liquidityNet, uint256 feeGrowthOutside0X128, uint256 feeGrowthOutside1X128, int56 tickCumulativeOutside, uint160 secondsPerLiquidityOutsideX128, uint32 secondsOutside, bool initialized)";
-  } else {
-    throw new Error("Invalid method name: " + methodName);
-  }
-}
-
-function getSubgraphEndpoint(chainId: ChainId): SubgraphEndpoint {
-  switch (chainId) {
-    case ChainId.MAINNET:
-    case ChainId.RINKEBY:
-      return { author: "uniswap", name: "uniswap-v3" };
-    case ChainId.OPTIMISM:
-      return { author: "ianlapham", name: "uniswap-optimism-dev" };
-    case ChainId.ARBITRUM_ONE:
-      return { author: "ianlapham", name: "arbitrum-minimal" };
-    default:
-      throw new Error("Unknown or Unsupported chain ID");
-  }
-}
-
-function ethCallView(
-  address: string,
-  chainId: ChainId,
-  method: string
-): string {
-  return Ethereum_Module.callContractView({
-    address: address,
-    method: method,
-    args: [],
-    connection: {
-      node: null,
-      networkNameOrChainId: getChainIdKey(chainId),
-    },
-  }).unwrap();
+  return fetchAllTicks(args.address, args.chainId);
 }
 
 function fetchPoolImmutables(
@@ -234,14 +168,18 @@ function fetchPoolState(address: string, chainId: ChainId): PoolState {
   };
 }
 
-function fetchPoolTicksSubgraph(address: string, chainId: ChainId): Tick[] {
+function fetchPoolTicksSubgraph(
+  address: string,
+  chainId: ChainId,
+  skip: i32 = 0
+): Tick[] {
   const endpoint: SubgraphEndpoint = getSubgraphEndpoint(chainId);
   const query: JSON.Obj = subgraphQuery({
     subgraphAuthor: endpoint.author,
     subgraphName: endpoint.name,
     query: `
       query {
-        ticks(first: 1000, skip: 0, where: { poolAddress: "${address}" }, orderBy: tickIdx) {
+        ticks(first: 1000, skip: ${skip}, where: { poolAddress: "${address}" }, orderBy: tickIdx) {
           tickIdx
           liquidityGross
           liquidityNet
@@ -266,25 +204,13 @@ function fetchPoolTicksSubgraph(address: string, chainId: ChainId): Tick[] {
     );
 }
 
-function subgraphQuery(args: QueryArgs): JSON.Obj {
-  const response = Subgraph_Module.querySubgraph({
-    subgraphAuthor: args.subgraphAuthor,
-    subgraphName: args.subgraphName,
-    query: args.query,
-  }).unwrap();
-
-  const json = JSON.parse(response);
-
-  if (!json.isObj) {
-    throw new Error(
-      "Subgraph response is not an object.\n" +
-        `Author: ${args.subgraphAuthor}\n` +
-        `Subgraph: ${args.subgraphName}\n` +
-        `Query: ${args.query}\n` +
-        `Response: ${response}`
-    );
+function fetchAllTicks(address: string, chainId: ChainId): Tick[] {
+  let skip: i32 = 0;
+  let ticks: Tick[] = fetchPoolTicksSubgraph(address, chainId, skip);
+  while (ticks.length % 1000 === 0) {
+    skip += 1000;
+    const query = fetchPoolTicksSubgraph(address, chainId, skip);
+    ticks = ticks.concat(query);
   }
-
-  const obj = json as JSON.Obj;
-  return obj.valueOf().get("data") as JSON.Obj;
+  return ticks;
 }
